pyproject.toml
[project]
name = "sb2solace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "orjson>=3.10.7",
    "pyrsolace>=0.1.3",
    "python-dotenv>=1.0.1",
]

[project.scripts]
main = "sb2solace.__main__:main"
sub = "sb2solace.__main__:sub"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
